+++ lib/src/utils/extensions/build_context_extensions.dart
@@ -3,5 +3,5 @@ import 'package:provider/provider.dart' show Provider;
 import 'package:syncx/src/notifier/base/base_notifier.dart' show BaseNotifier;
 
 extension BuildContextExtensions on BuildContext {
-  N get<N extends BaseNotifier>() => Provider.of<N>(this, listen: false);
+  N read<N extends BaseNotifier>() => Provider.of<N>(this, listen: false);
 }