+++ lib/syncx.dart
@@ -12,8 +12,6 @@
 /// See the package README and example for usage details.
 library SyncX;
 
-export 'package:provider/provider.dart' show ReadContext;
-
 export 'src/builders/async/async_notifier_builder.dart';
 export 'src/builders/async/async_notifier_consumer.dart';
 export 'src/builders/async/async_notifier_listener.dart';
@@ -21,5 +19,7 @@ export 'src/builders/base/base_notifier_builder.dart'
     show NotifierBuilder, NotifierConsumer, NotifierListener;
 export 'src/notifier/base/base_notifier.dart' show AsyncNotifier, Notifier;
 export 'src/register/notifier_register.dart';
+export 'src/utils/extensions/build_context_extensions.dart'
+    show BuildContextExtensions;
 export 'src/utils/models/async_state.dart';
 export 'src/utils/models/error_state.dart';